requirements.txt
requests
rich
python-dotenv
streamlit
plotly
pandas
